requirements.txt
python-dotenv
pytest
pandas